+++ Makefile
@@ -24,7 +24,7 @@ ASAN_FLAGS := -g -O0 -fsanitize=address -fno-omit-frame-pointer
 _SRCS := main.c network.c game_logic.c game_server.c game_types.c json_messages.c client.c util.c hash_table.c room.c timer.c
 SERVER_SRC := $(addprefix $(SERVER_DIR)/,$(_SRCS))
 
-_TEST_SRCS := test.c hash_table.c util.c timer.c
+_TEST_SRCS := test.c hash_table.c util.c timer.c test_util.c
 TEST_SRC := $(addprefix $(SERVER_DIR)/,$(_TEST_SRCS))
 
 # ==========================================
