+++ server/test_util.h
@@ -1,0 +1,8 @@
+#ifndef TEST_UTIL_H
+#define TEST_UTIL_H
+
+typedef void (*Test)(void);
+
+void sleep_with_log(unsigned int seconds);
+
+#endif // !TEST_UTIL_H
