+++ server/util_test.c
@@ -1,0 +1,12 @@
+#include "util.h"
+#include <assert.h>
+#include <stdlib.h>
+#include <string.h>
+
+static void test_generate_random_string(void);
+
+void test_generate_random_string(void) {
+  char *s = generate_random_string(7);
+  assert(strlen(s) == 7);
+  free(s);
+}
